refactor: rename `config_toml` to `config` and `config` to `environment`
