+++ src/cli/build.rs
@@ -10,14 +10,15 @@ use eyre::{eyre, WrapErr};
 use crate::{
     assets_sync,
     compiler::{self, all_trees_source},
-    config::{self, output_path, BuildMode},
-    config_toml, html_flake,
+    config,
+    environment::{self, output_path, BuildMode},
+    html_flake,
 };
 
 #[derive(clap::Args)]
 pub struct BuildCommand {
     /// Path to the configuration file (e.g., "kodama.toml").
-    #[arg(short, long, default_value_t = config_toml::DEFAULT_CONFIG_PATH.into())]
+    #[arg(short, long, default_value_t = config::DEFAULT_CONFIG_PATH.into())]
     config: String,
 }
 
@@ -27,14 +28,14 @@ pub fn build(command: &BuildCommand) -> eyre::Result<()> {
 }
 
 pub fn build_with(config: &str, mode: BuildMode) -> eyre::Result<()> {
-    config::init_environment(config.into(), mode)?;
+    environment::init_environment(config.into(), mode)?;
 
-    if !config::inline_css() {
+    if !environment::inline_css() {
         export_css_files().wrap_err("failed to export CSS")?;
     }
 
-    let root = config::root_dir();
-    let workspace = all_trees_source(&config::trees_dir())?;
+    let root = environment::root_dir();
+    let workspace = all_trees_source(&environment::trees_dir())?;
     compiler::compile(workspace).wrap_err_with(|| {
         eyre!(
             "failed to compile site `{}`",
@@ -66,8 +67,8 @@ fn export_css_file(css_content: &str, name: &str) -> eyre::Result<()> {
 /// Synchronize the assets directory [`config::assets_dir`] with the
 /// output directory [`config::output_dir()`].
 fn sync_assets_dir() -> eyre::Result<bool> {
-    let asset_dir = config::assets_dir();
-    let target = config::output_dir().join(asset_dir.file_name().unwrap());
+    let asset_dir = environment::assets_dir();
+    let target = environment::output_dir().join(asset_dir.file_name().unwrap());
 
     assets_sync::sync_assets(asset_dir, target)?;
     Ok(true)
